improved scraping and remove single target scrape
I think I finally finished fixing the scraping, so it only scrapes when
a request is made to the server and no longer on a set interval. But
implementing this I lost the single target scraping due to the way the
Prometheus register and collectors work (which I don't fully understand
yet).

So at this point in time the only scrape that is possible is to get all
the interface data. I need to read into the Prometheus and make it so
that the interfaces can be configured in a config file or through a URL query
parameter.

I do think there is still a better way then calling `.Collect(ch)` for
every metric, but at this time it will have to do.

I've also taken the liberty of using `prommod` so we can collect some
more application statistics from the `/app` endpoint.