further reworked the prometheus exporting
Moved each component into it's own collector. Before a hacky way was
designed to work around an issue. When tryong to introduce testing for
the collector, it was shown that the collector did not work correctly.

Will only merge this into amster after testing has been completed for
the exporter, but at first glance this seems to be the correct way to
work with prometheus exporters.