further implementation of exporting the namespace
I seem to have gotten everything to work with the current netns that the
program starts in, however I'm having some troubles connecting to other
netnses.

I think the netlink socket connects back to the running the
running thread of the program. This causes that the socket cannot connect
to a different netns than it is running in because of the program thread.

I'm not sure if this means that I need to spawn a new thread in the
other namespace to be able to access this, or maybe I'm just missing
something.