finished netns export implementation
Managed to figure it out, the functionality was in the library but I
just misunderstood it. The NetNS is not the ID of the netns as can be
deduced from `ip netns`, the NetNS is a file descriptor of the netns.

As it happens, all the netns have their own in `/var/run/netns`.