attempt at fixing concurrent map access lower in the Collect()
It could be that in theory the metics return early, while the scrape is still
runing. When Prometheus has a low scrape interval, this trigger concurrent map
access for the `ServiceCurveExporter` map in the `class.go`.

By removing the concurreny, we might slow down the exporter, but can confirm that
the concurrent access is causing issues. Otherwise this might be an issue deeper
down in the logic (and I have a bad feeling it is).