switch to viper for config and implement filters
On some system allowing all data to go through create very high cardinality labels.
By implementing a filter on the classes and qdiscs we can chose which tc objects are
parsed and send along. This allows for the filtering of fq_codel qdiscs for example
when using a hfsc setup.

Signed-off-by: Francis Begyn <francis.begyn@gmail.com>